Fix public-rest.spec.ts
